1) Added DEBUG define which enables some cheats, which are not    to be used in a release. Currently DEBUG enables the F1/2/3    cheats to give a player sparks, missiles, homing missiles, and    enables the player to call for a rescue while not touching    the ground. 2) Added many casts from size_t to int (and similar), to remove    warnings when compiling with windows. 3) Disabled F4 'bullet debug' shortcut, since it doesn't work    and it's not too useful anyway.
git-svn-id: svn+ssh://svn.code.sf.net/p/supertuxkart/code/trunk/supertuxkart@1333 178a84e3-b1eb-0310-8ba1-8eac791a3b58

+++ src/gui/config_display.cpp
@@ -225,7 +225,7 @@ void ConfigDisplay::getScreenModes()
             
             //Prevent use of very small resolutions
             const int minRes = 640;
-            m_sizes_size = m_sizes.size();
+            m_sizes_size = (int)m_sizes.size();
             
             for (int i = m_sizes_size-1; i >= 0; --i)
             {
@@ -240,7 +240,7 @@ void ConfigDisplay::getScreenModes()
     
     // search m_sizes for the current resolution
     m_sizes_index = -1;
-    m_sizes_size = m_sizes.size();
+    m_sizes_size = (int)m_sizes.size();
     for (int i = 0; i < m_sizes_size; i++)
     {
         if (m_sizes[i].first == user_config->m_width
